+++ javascript/src/deckgl-widget.js
@@ -1,28 +1,7 @@
+import {
+  createDeckGLProperties,
+  logVersions } from "./utils";
 import deckLayer from "./layer";
-
-// TODO: Use 'deckGLProperties' parameter in R
-function createDeckGLProperties(elementId, widgetData) {
-   return Object.assign({
-     mapboxApiAccessToken: widgetData.mapboxApiAccessToken || "",
-     mapStyle: widgetData.mapStyle || "",
-     container: elementId,
-     initialViewState: widgetData.initialViewState || createInitialViewState(widgetData),
-     views: widgetData.views || new deck.MapView(),
-     controller: true,
-     layers: [ ]
-   }, widgetData.properties);
-}
-
-// TODO: Create initial view state already in R
-function createInitialViewState(widgetData) {
-  return {
-    longitude: widgetData.longitude,
-    latitude: widgetData.latitude,
-    zoom: widgetData.zoom,
-    pitch: widgetData.pitch,
-    bearing: widgetData.bearing
-  };
-}
 
 export default function(widgetElement, width, height) {
   const widget = { };
@@ -43,7 +22,7 @@ export default function(widgetElement, width, height) {
 
   widget.renderValue = function(widgetData) {
     console.log(widgetData);
-    // widgetElement.innerHTML = "deckgl-widget";
+    logVersions();
 
     const deckGLProperties = createDeckGLProperties(widgetElement.id, widgetData);
     deckGL = new deck.DeckGL(deckGLProperties);
